uses reinforcement learning to teach an agent
Create an environment for the game (specifically, the CartPole game). We then define a neural network for the agent, with an input layer of size 4 (the number of observations in the CartPole game), two hidden layers of size 24, and an output layer of size 2 (the number of actions the agent can take). We use the Adam optimizer to train the neural network to predict the Q-values for each action given the current state.

We then use a loop to train the agent over 1000 episodes. In each episode, we start by resetting the environment and observing the initial state. We then use the neural network to choose an action based on the current state, and take that action to observe the next state and reward. We then update the Q-value for the current state and action using the Bellman equation, and train the neural network on this updated Q-value. We repeat this process for each time step in the episode, until the game is over (i.e. the pole falls over or the cart goes off the screen). Finally, we print out the number of timesteps it took for the game to end in each episode.